Impliment Heston Model of Stochastic Volatility (#30)
* Update training

Correct volatility of mean and volatility of volatility definition.

* Update training API

In order to be compatible with the Heston process the API of the training is updated. Pearson's correlation is used to get correlation between Weiner process of Volatility and Mean.

* Change in API reflected in example Notebook.

* Update model

Heston Model class added. Instantiate Heston Model class.

* Update model

Implement Heston Model

* Fix Heston model

Bugs fixed with Heston model